apt-mark error fixed

+++ aarch64/install.sh
@@ -65,7 +65,7 @@ env_check(){
         $PG install -y curl wget
     fi
     # Check if the system supports
-    wget  -nv --show-progress -O $TMP/screenfetch "https://raw.githubusercontent.com/KittyKatt/screenFetch/master/screenfetch-dev" 
+    [ ! -s $TMP/screenfetch ]&&wget  -nv --show-progress -O $TMP/screenfetch "https://raw.githubusercontent.com/KittyKatt/screenFetch/master/screenfetch-dev" 
     chmod +x $TMP/screenfetch
     OS=`$TMP/screenfetch -n |grep 'OS:'|awk '{print $3}'|tr 'A-Z' 'a-z'`
     if [[ -z "$OS" ]]; then
@@ -273,6 +273,7 @@ EOF
         echo "deb https://mirrors.aliyun.com/kubernetes/apt/ kubernetes-xenial main"|tee /etc/apt/sources.list.d/kubernetes.list
         log "[info]" "installing k8s"
         apt update
+        apt-mark unhold kubelet kubeadm kubectl kubernetes-cni
         apt install -y --allow-downgrades kubeadm=1.12.3-00 kubectl=1.12.3-00 kubelet=1.12.3-00 kubernetes-cni=0.6.0-00 
         apt-mark hold kubelet kubeadm kubectl kubernetes-cni
     }
@@ -493,3 +494,4 @@ case $1 in
         fi
         ;;
 esac
+sync
